Second draft, move to one class (I think that's what it's called). *Definitely* doesn't work.
