+++ src/main/java/com/ldtteam/structurize/blocks/bricks/BlockBricks.java
@@ -1,14 +1,14 @@
-package com.ldtteam.structurize.blocks.decorative;
+package com.ldtteam.structurize.blocks.bricks;
 
 import com.ldtteam.structurize.blocks.AbstractBlockStructurize;
 import net.minecraft.block.Block;
 import net.minecraft.block.Blocks;
 
-public class BlockBeigeBricks extends AbstractBlockStructurize<BlockBeigeBricks>
+public class BlockBricks extends AbstractBlockStructurize<BlockBricks>
 {
-    private static final String BLOCK_NAME = "blockbeigebricks";
+    private static final String BLOCK_NAME = "blockbricks";
 
-    public BlockBeigeBricks()
+    public BlockBricks()
     {
         super(Block.Properties.from(Blocks.BRICKS));
         setRegistryName(BLOCK_NAME);
